Rename unit tests implementations for conform with style guides.
