+++ Src/Shared/Managed-Src/Temporal.Util/internal/Validate.cs
