+++ Src/Test/SDK/Common/Temporal.TestUtil/internal/SerializableClass.cs
@@ -1,10 +1,25 @@
+using Xunit;
+
 namespace Temporal.TestUtil
 {
     internal class SerializableClass
     {
-        public static SerializableClass Default
+        public static SerializableClass CreateDefault()
         {
-            get { return new SerializableClass { Name = "Test", Value = 1 }; }
+            return new SerializableClass { Name = "Test", Value = 1 };
+        }
+
+        public static void AssertEqual(SerializableClass expected, SerializableClass actual)
+        {
+            if (expected == null)
+            {
+                Assert.Null(actual);
+                return;
+            }
+
+            Assert.NotNull(actual);
+            Assert.Equal(expected.Name, actual.Name);
+            Assert.Equal(expected.Value, actual.Value);
         }
 
         public string Name { get; set; }
